requirements.txt
opencv-python
matplotlib
python-dotenv